requirements.txt
# Microsoft Agent Framework (MAF)
agent-framework-core>=0.1.0
agent-framework-azure-ai>=0.1.0
agent-framework-orchestrations>=0.1.0

# Azure Identity
azure-identity>=1.15.0

# Azure AI Foundry SDK (for direct API access if needed)
azure-ai-projects>=2.0.0b3

# Azure AI Search (Phase 3 - RAG)
azure-search-documents>=11.7.0b2

# Azure Speech (Phase 2 - Voice Gateway)
azure-cognitiveservices-speech>=1.35.0

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0
rich>=13.0.0

# Data validation
pydantic>=2.5.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
